backend/requirements.txt
Django==6.0.1
django-environ==0.11.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-oauth-toolkit==2.3.0
drf-spectacular==0.29.0
django-storages==1.14.2
django-mongodb-backend==6.0.0
pymongo==4.16.0
uvicorn==0.40.0
gunicorn==25.0.1
whitenoise==6.11.0
